htmx_app/requirements.txt
fastapi
websockets
uvicorn[standard]
jinja